Add generic overlay component
